LPS-153356 Create new page-content plugin and move the Page contents code there
