+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/components/App.scss
@@ -54,4 +54,5 @@
 @import '../../plugins/comments/components/Plugin';
 @import '../../plugins/experience/components/Plugin';
 @import '../../plugins/fragments-widgets/components/Plugin';
+@import '../../plugins/page-content/components/Plugin';
 @import '../../plugins/page-design-options/components/Plugin';
