+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/plugins/browser/components/_Plugin.scss
@@ -1,4 +1,3 @@
 @import 'BrowserSidebar';
 
-@import 'contents/components/Plugin';
 @import 'page-structure/components/Plugin';
