+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/plugins/page-content/components/ContentsSidebar.js
@@ -14,14 +14,15 @@
 
 import React, {useMemo} from 'react';
 
-import {EDITABLE_FRAGMENT_ENTRY_PROCESSOR} from '../../../../../app/config/constants/editableFragmentEntryProcessor';
-import {EDITABLE_TYPES} from '../../../../../app/config/constants/editableTypes';
-import {useSelector} from '../../../../../app/contexts/StoreContext';
-import selectLanguageId from '../../../../../app/selectors/selectLanguageId';
-import {selectPageContents} from '../../../../../app/selectors/selectPageContents';
-import isMapped from '../../../../../app/utils/editable-value/isMapped';
-import {getEditableLocalizedValue} from '../../../../../app/utils/getEditableLocalizedValue';
-import SidebarPanelContent from '../../../../../common/components/SidebarPanelContent';
+import {EDITABLE_FRAGMENT_ENTRY_PROCESSOR} from '../../../app/config/constants/editableFragmentEntryProcessor';
+import {EDITABLE_TYPES} from '../../../app/config/constants/editableTypes';
+import {useSelector} from '../../../app/contexts/StoreContext';
+import selectLanguageId from '../../../app/selectors/selectLanguageId';
+import {selectPageContents} from '../../../app/selectors/selectPageContents';
+import isMapped from '../../../app/utils/editable-value/isMapped';
+import {getEditableLocalizedValue} from '../../../app/utils/getEditableLocalizedValue';
+import SidebarPanelContent from '../../../common/components/SidebarPanelContent';
+import SidebarPanelHeader from '../../../common/components/SidebarPanelHeader';
 import NoPageContents from './NoPageContents';
 import PageContents from './PageContents';
 
@@ -124,6 +125,11 @@ export default function ContentsSidebar() {
 
 	return (
 		<>
+			{Liferay.FeatureFlags['LPS-153452'] && (
+				<SidebarPanelHeader>
+					{Liferay.Language.get('page-content')}
+				</SidebarPanelHeader>
+			)}
 			<SidebarPanelContent
 				className="page-editor__page-contents"
 				padded={false}
