+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/plugins/page-content/components/PageContents.js
@@ -15,7 +15,7 @@
 import PropTypes from 'prop-types';
 import React, {useMemo, useState} from 'react';
 
-import {CONTENT_TYPE_LABELS} from '../../../../../app/config/constants/contentTypeLabels';
+import {CONTENT_TYPE_LABELS} from '../../../app/config/constants/contentTypeLabels';
 import ContentFilter from './ContentFilter';
 import ContentList from './ContentList';
